valac: change so that vala2c output target instead of variable
this way, we can use the target to sort the various output files. while we are at it, drop the _NAME suffix from some of the properties to match the existing target property patterns (e.g. TARGET_FILE is the full path while TARGET_FILE_NAME is just the file name)